// src/domain/enums/ComplianceCheckType.ts
export enum ComplianceCheckType {
  VAT_VALIDATION = "VAT_VALIDATION",
  SUPPLIER_VERIFICATION = "SUPPLIER_VERIFICATION",
  DOCUMENT_VALIDITY = "DOCUMENT_VALIDITY",
  AMOUNT_CHECK = "AMOUNT_CHECK",
  DUPLICATE_CHECK = "DUPLICATE_CHECK",
  TAX_ID_VALIDATION = "TAX_ID_VALIDATION",
  SANCTIONS_SCREENING = "SANCTIONS_SCREENING",
  PEP_SCREENING = "PEP_SCREENING",
  AML_CHECK = "AML_CHECK",
  KYC_VERIFICATION = "KYC_VERIFICATION",
  REGULATORY_COMPLIANCE = "REGULATORY_COMPLIANCE",
  DATA_PROTECTION = "DATA_PROTECTION",
  FRAUD_DETECTION = "FRAUD_DETECTION",
}